requirements.txt
Flask
gunicorn
RPi.GPIO
